pyproject.toml
[tool.poetry]
name = "data-rights-protocol"
version = "0"
description = "datarightsprotocol.org"
authors = ["Ryan Rix <ryan.rix.consultant@consumer.org>"]
license = "Apache 2"

[tool.poetry.dependencies]
python = "^3.9"
pyjwt = "^2.3.0"
uuid = "^1.30"
hypothesis = "^6.35.0"

[tool.poetry.dependencies.pydantic]
extras = ["email"]
version = "^1.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
